# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Task Center API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  urgent: 'urgent',
} as const;

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  assigneeName?: string | null;
  /** @nullable */
  sourceMessageId?: number | null;
  /** @nullable */
  dueDate?: string | null;
  tags?: string[];
  createdAt: string;
  updatedAt: string;
}

export type TaskInputStatus = typeof TaskInputStatus[keyof typeof TaskInputStatus];


export const TaskInputStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type TaskInputPriority = typeof TaskInputPriority[keyof typeof TaskInputPriority];


export const TaskInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  urgent: 'urgent',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  status?: TaskInputStatus;
  priority: TaskInputPriority;
  assigneeId?: number;
  dueDate?: string;
  tags?: string[];
}

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type TaskUpdatePriority = typeof TaskUpdatePriority[keyof typeof TaskUpdatePriority];


export const TaskUpdatePriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  urgent: 'urgent',
} as const;

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  status?: TaskUpdateStatus;
  priority?: TaskUpdatePriority;
  assigneeId?: number;
  dueDate?: string;
  tags?: string[];
}

export interface TaskAssignment {
  assigneeId: number;
}

export interface TeamMember {
  id: number;
  name: string;
  role: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface TeamMemberInput {
  /** @minLength 1 */
  name: string;
  role: string;
  phone?: string;
  email?: string;
  avatarUrl?: string;
}

export interface TeamMemberUpdate {
  name?: string;
  role?: string;
  phone?: string;
  email?: string;
  avatarUrl?: string;
}

export interface WhatsAppMessage {
  id: number;
  from: string;
  /** @nullable */
  senderName?: string | null;
  body: string;
  timestamp: string;
  processed: boolean;
  /** @nullable */
  detectedIntent?: string | null;
  /** @nullable */
  taskId?: number | null;
  createdAt: string;
}

export interface MessageProcessResult {
  intent: string;
  taskCreated: boolean;
  task?: Task;
}

export interface SendMessageInput {
  to: string;
  message: string;
}

export interface SendMessageResult {
  success: boolean;
  /** @nullable */
  messageId?: string | null;
  /** @nullable */
  error?: string | null;
}

export type WebhookPayloadEntryItem = { [key: string]: unknown };

export interface WebhookPayload {
  object?: string;
  entry?: WebhookPayloadEntryItem[];
}

export type DocumentStatus = typeof DocumentStatus[keyof typeof DocumentStatus];


export const DocumentStatus = {
  pending: 'pending',
  auditing: 'auditing',
  audited: 'audited',
  error: 'error',
} as const;

export interface Document {
  id: number;
  filename: string;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  storagePath?: string | null;
  /** @nullable */
  mimeType?: string | null;
  /** @nullable */
  fileSize?: number | null;
  status: DocumentStatus;
  /** @nullable */
  auditSummary?: string | null;
  auditIssues?: string[];
  /** @nullable */
  auditScore?: number | null;
  /** @nullable */
  taskId?: number | null;
  /** @nullable */
  uploadedBy?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface DocumentInput {
  filename: string;
  fileUrl?: string;
  storagePath?: string;
  mimeType?: string;
  fileSize?: number;
  taskId?: number;
  uploadedBy?: string;
}

export interface UploadUrlRequest {
  filename: string;
  mimeType: string;
}

export interface UploadUrlResponse {
  uploadUrl: string;
  publicUrl: string;
  path: string;
}

export interface DashboardStats {
  totalTasks: number;
  openTasks: number;
  completedTasks: number;
  urgentTasks: number;
  totalMessages: number;
  pendingMessages: number;
  totalDocuments: number;
  auditedDocuments: number;
  teamSize: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  task_created: 'task_created',
  task_updated: 'task_updated',
  task_assigned: 'task_assigned',
  message_received: 'message_received',
  document_audited: 'document_audited',
  member_added: 'member_added',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  /** @nullable */
  entityId?: number | null;
  createdAt: string;
}

export type ListTasksParams = {
status?: string;
assigneeId?: number;
};

export type ListMessagesParams = {
processed?: boolean;
};

export type VerifyWhatsAppWebhookParams = {
'hub.mode'?: string;
'hub.verify_token'?: string;
'hub.challenge'?: string;
};