# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Task Center API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List received WhatsApp messages
      parameters:
        - in: query
          name: processed
          schema:
            type: boolean
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WhatsAppMessage"
  /messages/{id}/process:
    post:
      operationId: processMessage
      tags: [messages]
      summary: Re-run AI intent detection on a message and create task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Processing result with detected intent and created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageProcessResult"
  /messages/send:
    post:
      operationId: sendWhatsAppMessage
      tags: [messages]
      summary: Send a WhatsApp notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageInput"
      responses:
        "200":
          description: Send result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendMessageResult"

  # ── WhatsApp Webhook ─────────────────────────────────────────────────────────
  /webhook/whatsapp:
    get:
      operationId: verifyWhatsAppWebhook
      tags: [messages]
      summary: WhatsApp webhook verification (GET challenge)
      parameters:
        - in: query
          name: hub.mode
          schema:
            type: string
        - in: query
          name: hub.verify_token
          schema:
            type: string
        - in: query
          name: hub.challenge
          schema:
            type: string
      responses:
        "200":
          description: Challenge echoed back
    post:
      operationId: receiveWhatsAppWebhook
      tags: [messages]
      summary: Receive incoming WhatsApp message events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookPayload"
      responses:
        "200":
          description: Acknowledged

  # ── Documents ────────────────────────────────────────────────────────────────
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all audited documents
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a document for AI audit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created document record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /documents/{id}/audit:
    post:
      operationId: auditDocument
      tags: [documents]
      summary: Run AI audit on an uploaded document
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Audit result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  # ── Upload presigned URL (Supabase Storage) ──────────────────────────────────
  /documents/upload-url:
    post:
      operationId: getUploadUrl
      tags: [documents]
      summary: Get a presigned upload URL from Supabase Storage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"

  # ── Dashboard / Analytics ────────────────────────────────────────────────────
...
components:
  schemas:
...
    Task:
      type: object
      required: [id, title, status, priority, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assigneeId:
          type: ["integer", "null"]
        assigneeName:
          type: ["string", "null"]
        sourceMessageId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
    WhatsAppMessage:
      type: object
      required: [id, from, body, timestamp, processed, createdAt]
      properties:
        id:
          type: integer
        from:
          type: string
        senderName:
          type: ["string", "null"]
        body:
          type: string
        timestamp:
          type: string
        processed:
          type: boolean
        detectedIntent:
          type: ["string", "null"]
        taskId:
          type: ["integer", "null"]
        createdAt:
          type: string
    MessageProcessResult:
      type: object
      required: [intent, taskCreated]
      properties:
        intent:
          type: string
        taskCreated:
          type: boolean
        task:
          $ref: "#/components/schemas/Task"
    SendMessageInput:
      type: object
      required: [to, message]
      properties:
        to:
          type: string
        message:
          type: string
    SendMessageResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        messageId:
          type: ["string", "null"]
        error:
          type: ["string", "null"]
    WebhookPayload:
      type: object
      properties:
        object:
          type: string
        entry:
          type: array
          items:
            type: object

    # ── Document ──────────────────────────────────────────────────────────────
    Document:
      type: object
      required: [id, filename, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        filename:
          type: string
        fileUrl:
          type: ["string", "null"]
        storagePath:
          type: ["string", "null"]
        mimeType:
          type: ["string", "null"]
        fileSize:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, auditing, audited, error]
        auditSummary:
          type: ["string", "null"]
        auditIssues:
          type: array
          items:
            type: string
        auditScore:
          type: ["integer", "null"]
        taskId:
          type: ["integer", "null"]
        uploadedBy:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    DocumentInput:
      type: object
      required: [filename]
      properties:
        filename:
          type: string
        fileUrl:
          type: string
        storagePath:
          type: string
        mimeType:
          type: string
        fileSize:
          type: integer
        taskId:
          type: integer
        uploadedBy:
          type: string
    UploadUrlRequest:
      type: object
      required: [filename, mimeType]
      properties:
        filename:
          type: string
        mimeType:
          type: string
    UploadUrlResponse:
      type: object
      required: [uploadUrl, publicUrl, path]
      properties:
        uploadUrl:
          type: string
        publicUrl:
          type: string
        path:
          type: string

    # ── Dashboard ─────────────────────────────────────────────────────────────
...