# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Task Center API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentStatus } from './documentStatus';

export interface Document {
  id: number;
  filename: string;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  storagePath?: string | null;
  /** @nullable */
  mimeType?: string | null;
  /** @nullable */
  fileSize?: number | null;
  status: DocumentStatus;
  /** @nullable */
  auditSummary?: string | null;
  auditIssues?: string[];
  /** @nullable */
  auditScore?: number | null;
  /** @nullable */
  taskId?: number | null;
  /** @nullable */
  uploadedBy?: string | null;
  createdAt: string;
  updatedAt: string;
}