# lib/api-zod/src/generated/types/sendMessageResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Task Center API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SendMessageResult {
  success: boolean;
  /** @nullable */
  messageId?: string | null;
  /** @nullable */
  error?: string | null;
}